refactor: input abstract
