+++ src/app/shared/components/ui/card/card-product-details/card-product-details.component.ts
@@ -26,7 +26,7 @@ import { ProductService } from '@services/product.service';
 import { WishlistService } from '@services/wishlist.service';
 import { ButtonCtaLargeComponent } from '@shared/components/ui/buttons/button-cta-large/button-cta-large.component';
 import { ButtonComponent } from '@shared/components/ui/buttons/button/button.component';
-import { NumberStepFieldComponent } from '@shared/components/ui/form/number-step-field/number-step-field.component';
+import { NumberStepFieldComponent } from '@shared/components/ui/form/field/number-step-field/number-step-field.component';
 import { PriceDiscountPercentageComponent } from '@shared/components/ui/price/price-discount-percentage/price-discount-percentage.component';
 import { PriceComponent } from '@shared/components/ui/price/price/price.component';
 import { RatingComponent } from '@shared/components/ui/rating/rating.component';
