+++ src/app/shared/components/ui/form/field/checkbox/checkbox-color-field/checkbox-color-field.component.html
