+++ src/app/shared/components/ui/form/field/checkbox/checkbox-color-field/checkbox-color-field.component.ts
@@ -1,23 +1,24 @@
-import { FormInputAbstract } from '@abstract/form-input.abstract';
+import { FormInputAbstract } from '@abstract/form-input/form-input.abstract';
 import { Component, InputSignal, OutputEmitterRef, forwardRef, input, output } from '@angular/core';
 import { FormsModule, NG_VALUE_ACCESSOR, ReactiveFormsModule } from '@angular/forms';
 import { CheckboxChangeEvent, CheckboxModule } from 'primeng/checkbox';
 
 @Component({
-	selector: 'app-checkbox-field',
+	selector: 'app-checkbox-color-field',
 	imports: [CheckboxModule, FormsModule, ReactiveFormsModule],
-	templateUrl: './checkbox-field.component.html',
-	styleUrl: './checkbox-field.component.scss',
+	templateUrl: './checkbox-color-field.component.html',
+	styleUrl: './checkbox-color-field.component.scss',
 	providers: [
 		{
 			provide: NG_VALUE_ACCESSOR,
-			useExisting: forwardRef(() => CheckboxFieldComponent),
+			useExisting: forwardRef(() => CheckboxColorFieldComponent),
 			multi: true,
 		},
 	],
 })
-export class CheckboxFieldComponent extends FormInputAbstract<string[]> {
-	public checkboxValue: InputSignal<string> = input.required<string>();
+export class CheckboxColorFieldComponent extends FormInputAbstract<string[], CheckboxChangeEvent> {
+	public checkboxValue: InputSignal<string | number> = input.required<string | number>();
+	public colorClassCss: InputSignal<string> = input.required<string>();
 	public changed: OutputEmitterRef<void> = output();
 
 	public override onInputChange(event: CheckboxChangeEvent): void {
