+++ src/app/shared/components/ui/form/field/checkbox/checkbox-field/checkbox-field.component.scss
