+++ src/app/shared/components/ui/form/field/checkbox/checkbox-field/checkbox-field.component.ts
@@ -1,26 +1,23 @@
-import { FormInputAbstract } from '@abstract/form-input.abstract';
+import { FormInputWithLabelAbstract } from '@abstract/form-input/form-input-with-label.abstract';
 import { Component, InputSignal, OutputEmitterRef, forwardRef, input, output } from '@angular/core';
 import { FormsModule, NG_VALUE_ACCESSOR, ReactiveFormsModule } from '@angular/forms';
 import { CheckboxChangeEvent, CheckboxModule } from 'primeng/checkbox';
 
 @Component({
-	selector: 'app-checkbox-color-field',
+	selector: 'app-checkbox-field',
 	imports: [CheckboxModule, FormsModule, ReactiveFormsModule],
-	templateUrl: './checkbox-color-field.component.html',
-	styleUrl: './checkbox-color-field.component.scss',
+	templateUrl: './checkbox-field.component.html',
+	styleUrl: './checkbox-field.component.scss',
 	providers: [
 		{
 			provide: NG_VALUE_ACCESSOR,
-			useExisting: forwardRef(() => CheckboxColorFieldComponent),
+			useExisting: forwardRef(() => CheckboxFieldComponent),
 			multi: true,
 		},
 	],
 })
-export class CheckboxColorFieldComponent extends FormInputAbstract<string[]> {
-	public override labelFor: InputSignal<string> = input<string>('');
-	public override label: InputSignal<string> = input<string>('');
-	public checkboxValue: InputSignal<string | number> = input.required<string | number>();
-	public colorClassCss: InputSignal<string> = input.required<string>();
+export class CheckboxFieldComponent extends FormInputWithLabelAbstract<string[], CheckboxChangeEvent> {
+	public checkboxValue: InputSignal<string> = input.required<string>();
 	public changed: OutputEmitterRef<void> = output();
 
 	public override onInputChange(event: CheckboxChangeEvent): void {
