+++ src/app/shared/components/ui/form/field/mask-field/mask-field.component.scss
