+++ src/app/shared/components/ui/form/field/number-step-field/number-step-field.component.scss
