+++ src/app/shared/components/ui/form/field/password-field/password-field.component.scss
