+++ src/app/shared/components/ui/form/field/rating/rating.component.scss
