+++ src/app/shared/components/ui/form/field/search-field/search-field.component.scss
