+++ src/app/shared/components/ui/form/field/slider-field/slider-field.component.scss
