+++ src/app/shared/components/ui/form/field/textarea-field/textarea-field.component.scss
