+++ src/app/shared/components/ui/header/header/header.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { SearchFieldComponent } from '@shared/components/ui/form/search-field/search-field.component';
+import { SearchFieldComponent } from '@shared/components/ui/form/field/search-field/search-field.component';
 import { HeaderNavigationComponent } from '@shared/components/ui/header/header-navigation/header-navigation.component';
 import { HeaderPrimaryNavigationComponent } from '@shared/components/ui/header/header-primary-navigation/header-primary-navigation.component';
 import { HeaderTopbarComponent } from '@shared/components/ui/header/header-topbar/header-topbar.component';
