+++ src/app/shared/components/ui/tabs/tabs/tabs.component.ts
@@ -12,8 +12,8 @@ import { CommentService } from '@services/comment.service';
 import { ButtonFormComponent } from '@shared/components/ui/buttons/button-form/button-form.component';
 import { ButtonLargeComponent } from '@shared/components/ui/buttons/button-large/button-large.component';
 import { CardCommentComponent } from '@shared/components/ui/card/card-comment/card-comment.component';
-import { RatingComponent } from '@shared/components/ui/form/rating/rating.component';
-import { TextareaFieldComponent } from '@shared/components/ui/form/textarea-field/textarea-field.component';
+import { RatingComponent } from '@shared/components/ui/form/field/rating/rating.component';
+import { TextareaFieldComponent } from '@shared/components/ui/form/field/textarea-field/textarea-field.component';
 import { TabLiComponent } from '@shared/components/ui/tabs/tab-li/tab-li.component';
 import { TablistComponent } from '@shared/components/ui/tabs/tablist/tablist.component';
 import { TabsModule } from 'primeng/tabs';
